requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
aniso8601==9.0.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
dnspython==2.4.2
exceptiongroup==1.1.3
filelock==3.12.4
FireWorks==2.0.3
fireworks-ai==0.4.1
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
flask-paginate==2022.1.8
flask-restplus==0.13.0
flask-restx==1.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
frozenlist==1.4.0
fsspec==2023.9.0
Flask-JWT-Extended==4.5.3
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
httpx-sse==0.3.1
huggingface-hub==0.17.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
langchain==0.0.320
langsmith==0.0.49
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
monty==2023.9.5
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.6
numpy
openai==0.28.0
opencv-python==4.8.0.76
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
pytesseract==0.3.10
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sympy
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2

